# arrayToQPath: use QByteArray as backing store