this avoids:
1) allocation and zero-ing of bytearray
2) copy of bytearray to QByteArray

Using the benchmarking script in #1817,

test | time (s) / master | time (s) / this PR
-- | -- | --
pairs | 0.394 | 0.360
array | 0.394 | 0.365