@j9ac9k 2021-08-07
9% improvement for "free"; nice!  ...also getting away from if statements based on Qt binding versions; this diff is fantastic.

Regarding the two points, does this mean we likely have some memory consumption benefits as well (at least temporarily until the function returns?).