@pijyoi 2021-08-07
>does this mean we likely have some memory consumption benefits as well

Yes, there is one less memory allocation.

In the pre-#1796 code: (3x allocs + 2x memcpy for PySide; 2x allocs + 1x memcpy for PyQt)
1) one ndarray.empty() (alloc-only)
2) one conversion to bytes (alloc + memcpy) (for PySide codepath)
3) one QByteArray (alloc + memcpy)

In master: (2x allocs + 1x zero + 1x memcpy)
1) one bytearray (alloc + zero)
2) one QByteArray (alloc + memcpy)

This PR: (1x alloc)
1) one QByteArray (alloc-only)