@j9ac9k 2021-08-07
> > does this mean we likely have some memory consumption benefits as well
> 
> Yes, there is one less memory allocation.
> 
> In the pre-#1796 code: (3x allocs + 2x memcpy for PySide; 2x allocs + 1x memcpy for PyQt)
> 
> 1. one ndarray.empty() (alloc-only)
> 2. one conversion to bytes (alloc + memcpy) (for PySide codepath)
> 3. one QByteArray (alloc + memcpy)
> 
> In master: (2x allocs + 1x zero + 1x memcpy)
> 
> 1. one bytearray (alloc + zero)
> 2. one QByteArray (alloc + memcpy)
> 
> This PR: (1x alloc)
> 
> 1. one QByteArray (alloc-only)

Oh wow, as part of an IEEE paper we submitted some time ago (which did not get accepted) we did some [benchmarking of plot refresh rates vs. number of points in a line](https://github.com/pyqtgraph/IEEEVIS2021-pyqtgraph/blob/main/figures/lineBenchmark.pdf); ....this was pre #1796; we will be trying to resubmit that paper elsewhere (let me know if you're interested in joining that effort); but I'm going to be curious as to the impact of not only the speed boost but if memory consumption will play a role as well (basically, will _really_ large line-plots have significant gains...  If/when I rerun those benchmarks, I'll be sure to update you